extra scraping class added
